Move SDK job classes into own package
