+++ src/us/kbase/narrativejobservice/sdkjobs/DockerRunner.java
@@ -1,4 +1,4 @@
-package us.kbase.narrativejobservice;
+package us.kbase.narrativejobservice.sdkjobs;
 
 import java.io.BufferedReader;
 import java.io.File;
